+++ src/neurotorch/utils/visualise.py
@@ -1,0 +1,35 @@
+import numpy as np
+from matplotlib import pyplot as plt
+
+
+def plot_confusion_matrix(cm, classes, ):
+	import matplotlib.pyplot as plt
+	import itertools
+	
+	plt.imshow(cm, interpolation='nearest', cmap=plt.cm.Blues)
+	plt.title('Confusion matrix')
+	plt.colorbar()
+	tick_marks = np.arange(len(classes))
+	plt.xticks(tick_marks, classes, rotation=45)
+	plt.yticks(tick_marks, classes)
+	
+	fmt = 'd'
+	thresh = cm.max() / 2.
+	for i, j in itertools.product(range(cm.shape[0]), range(cm.shape[1])):
+		plt.text(
+			j, i,
+			format(cm[i, j], fmt),
+			horizontalalignment="center",
+			color="white" if cm[i, j] > thresh else "black"
+		)
+	
+	plt.ylabel('True label')
+	plt.xlabel('Predicted label')
+	plt.tight_layout()
+	plt.show()
+
+
+def legend_without_duplicate_labels_(ax: plt.Axes):
+	handles, labels = ax.get_legend_handles_labels()
+	unique = [(h, l) for i, (h, l) in enumerate(zip(handles, labels)) if l not in labels[:i]]
+	ax.legend(*zip(*unique))
